Added one more condition in board::move
If you refer to the modified code block in board.cpp, the only difference is that under handling promotion block, I added one more condition to check if the move type is en-passant. This avoid the issue where whenever a pawn move (weather capture or simply just move), the space object that the dest points to will always be deleted.